structure into resuable component
