+++ src/pages/aurora-site/course-listing/course-listing-page.jsx
@@ -1,4 +1,4 @@
-import CourseListingPage from "@/components/course-listing/course-listing";
+import CourseListingPage from "@/components/course-listing/course-listing-page";
 
 export default function CourseListing() {
   return <CourseListingPage />;
